Include JDBCStoreAutoConfiguration in GeoServerCatalogConfig
Include JDBCStoreAutoConfiguration in GeoServerCatalogConfig overriding
the bean definitions in
gs-jdbcscore-<version>.jar!/applicationContext.xml with a java
configuration in order to:
- share the java.sql.DataSource with jdbcconfig
- configure through spring properties (application.yml, etc)
- use a SimpleResourceCache with no data directory
