+++ services/catalog/pom.xml
@@ -12,8 +12,8 @@
   <packaging>jar</packaging>
   <dependencies>
     <dependency>
-      <groupId>org.geoserver</groupId>
-      <artifactId>gs-main</artifactId>
+      <groupId>org.geoserver.cloud</groupId>
+      <artifactId>gs-cloud-common</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
